requirements.txt
attrs==19.3.0
autopep8==1.5
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chalice==1.12.0
chardet==3.0.4
Click==7.0
docutils==0.15.2
enum-compat==0.0.3
filetype==1.0.5
idna==2.8
jmespath==0.9.4
Pillow==7.0.0
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.10.5
python-resize-image==1.1.19
requests==2.22.0
s3transfer==0.3.2
six==1.14.0
urllib3==1.25.8
